chore: workaround for Espressif's misuse of CMake (espressif/esp-idf#17773)

+++ include/ulf/mx1bin/encoder.hpp
@@ -66,7 +66,9 @@ struct Encoder {
 
   void addEOT() { *_iter++ = eot; }
 
-  size_t difference() { return std::distance(_begin, _iter); }
+  size_t difference() {
+    return static_cast<size_t>(std::distance(_begin, _iter));
+  }
 
   /// Encode uint8
   ///
